refactor(parser): split stmt/class.rs into cohesive submodules (§7-8)
Split the 2192-line parser/stmt/class.rs (over the 500-line limit) into a
facade plus six themed submodules under src/parser/stmt/class/, all under
500 lines. class.rs is entirely free parser functions addressed as
class::<name> (19 call sites); the facade re-exports each submodule
preserving its original visibility, so no external caller changes.
stmt/mod.rs is untouched.

- class.rs (41)          - facade: use lines + re-exports
- role_decl.rs (465)     - role-param parsing + role_decl
- class_decl.rs (458)    - class/augment/anon class decls + body + helpers
- package_decl.rs (405)  - package/unit-module/proto + export-clash checks
- token_body.rs (394)    - raw braced body / token regex injection
- attr_checks.rs (304)   - attr-twigil / no-self validation walkers
- grammar_module.rs (206)- does/trusts/token/grammar/module decls

Pure relocation. Moved fns are pub(crate) in submodules and re-exported at
their original visibility; cross-submodule private helpers bumped to
pub(crate). Body references to stmt-level siblings gained one super:: level.
The fn multiset is byte-identical. No behavior change.
