+++ src/parser/stmt/class.rs
@@ -1,2192 +1,41 @@
-use super::super::expr::expression;
-use super::super::helpers::{skip_balanced_parens, ws, ws1};
-use super::super::parse_result::{PError, PResult, opt_char, parse_char, take_while1};
-use super::super::primary::regex::scan_to_delim;
-use super::super::primary::var::is_pseudo_package;
+//! Class / role / grammar / module / package declaration parsers.
+//!
+//! This module is a facade: the parsing logic is split across themed
+//! submodules and re-exported here, preserving each function's original
+//! visibility so external callers (`class::class_decl`, etc.) keep working.
 
-use crate::ast::{Expr, ParamDef, Stmt};
-use crate::symbol::Symbol;
-use crate::token_kind::TokenKind;
-use crate::value::Value;
+mod attr_checks;
+mod class_decl;
+mod grammar_module;
+mod package_decl;
+mod role_decl;
+mod token_body;
 
-use super::{block, ident, keyword, qualified_ident};
+// Shared attribute / body validation helpers used across submodules.
+pub(crate) use attr_checks::{
+    null_regex_error, reject_no_self_in_attr_where, reject_no_self_in_subs,
+    reject_no_twigil_attr_at_body_level, stmt_also_is_parent, stmt_is_also_is_rw,
+};
 
-use super::sub::{parse_indirect_decl_name, parse_single_param, parse_sub_name};
-use super::{parse_param_list, parse_sub_traits};
+// Shared token / braced-body helpers used across submodules.
+pub(crate) use token_body::{
+    consume_raw_braced_body, inject_implicit_rule_ws, inject_separator_ws, normalize_token_pattern,
+    parse_raw_braced_regex_body, parse_token_like_name,
+};
 
-fn role_type_param_constraint_is_known(type_name: &str) -> bool {
-    let stripped = type_name
-        .strip_suffix(":D")
-        .or_else(|| type_name.strip_suffix(":U"))
-        .or_else(|| type_name.strip_suffix(":_"))
-        .unwrap_or(type_name);
-    let base = stripped
-        .split_once('[')
-        .map(|(head, _)| head)
-        .unwrap_or(stripped)
-        .split_once('(')
-        .map(|(head, _)| head)
-        .unwrap_or(stripped);
-    crate::runtime::utils::is_known_type_constraint(base)
-        || super::simple::is_user_declared_type(base)
-}
+// Shared declarator helpers used across submodules.
+pub(crate) use class_decl::{
+    meta_setter_stmt, parse_declarator_traits, parse_optional_bracket_suffix,
+};
+pub(crate) use package_decl::{
+    check_pseudo_package_in_decl, export_name_clash_error, extract_exported_subs,
+    find_export_name_clash,
+};
+pub(crate) use role_decl::skip_optional_role_args;
 
-fn parse_declarator_traits(input: &str) -> PResult<'_, Vec<(String, Value)>> {
-    let mut traits = Vec::new();
-    let (mut rest, _) = ws(input)?;
-    loop {
-        if !rest.starts_with(':') || rest.starts_with("::") {
-            break;
-        }
-        rest = &rest[1..];
-        let (r, name) = take_while1(rest, |c: char| c.is_alphanumeric() || c == '_' || c == '-')?;
-        rest = r;
-        let mut value = Value::Bool(true);
-        if let Some(inner) = rest.strip_prefix('<') {
-            if let Some(end) = inner.find('>') {
-                value = Value::str(inner[..end].to_string());
-                rest = &inner[end + 1..];
-            } else {
-                return Err(PError::expected("closing '>' in trait value"));
-            }
-        } else if rest.starts_with('(') {
-            let after = skip_balanced_parens(rest);
-            let body = &rest[1..rest.len() - after.len() - 1];
-            value = Value::str(body.trim().to_string());
-            rest = after;
-        }
-        traits.push((name.to_string(), value));
-        let (r, _) = ws(rest)?;
-        rest = r;
-    }
-    Ok((rest, traits))
-}
-
-fn parse_optional_bracket_suffix(input: &str) -> PResult<'_, String> {
-    if !input.starts_with('[') {
-        return Ok((input, String::new()));
-    }
-    let mut depth = 0u32;
-    let mut end = None;
-    for (i, ch) in input.char_indices() {
-        if ch == '[' {
-            depth += 1;
-        } else if ch == ']' {
-            depth = depth.saturating_sub(1);
-            if depth == 0 {
-                end = Some(i + 1);
-                break;
-            }
-        }
-    }
-    let end = end.ok_or_else(|| PError::expected("closing ']'"))?;
-    Ok((&input[end..], input[..end].to_string()))
-}
-
-fn split_role_param_parts(input: &str) -> Vec<&str> {
-    let mut parts = Vec::new();
-    let mut start = 0usize;
-    let mut paren = 0u32;
-    let mut bracket = 0u32;
-    let mut brace = 0u32;
-    let mut quote: Option<char> = None;
-    let mut escaped = false;
-    for (idx, ch) in input.char_indices() {
-        if let Some(q) = quote {
-            if escaped {
-                escaped = false;
-                continue;
-            }
-            if ch == '\\' && q == '"' {
-                escaped = true;
-                continue;
-            }
-            if ch == q {
-                quote = None;
-            }
-            continue;
-        }
-        match ch {
-            '\'' | '"' => quote = Some(ch),
-            '(' => paren += 1,
-            ')' => paren = paren.saturating_sub(1),
-            '[' => bracket += 1,
-            ']' => bracket = bracket.saturating_sub(1),
-            '{' => brace += 1,
-            '}' => brace = brace.saturating_sub(1),
-            ',' if paren == 0 && bracket == 0 && brace == 0 => {
-                parts.push(input[start..idx].trim());
-                start = idx + ch.len_utf8();
-            }
-            _ => {}
-        }
-    }
-    parts.push(input[start..].trim());
-    parts
-}
-
-fn meta_setter_stmt(type_name: &str, key: &str, value: Value) -> Stmt {
-    Stmt::Expr(Expr::Call {
-        name: Symbol::intern("__MUTSU_SET_META__"),
-        args: vec![
-            Expr::Literal(Value::str(type_name.to_string())),
-            Expr::Literal(Value::str(key.to_string())),
-            Expr::Literal(value),
-        ],
-    })
-}
-
-fn expr_uses_attr_twigil(expr: &Expr) -> bool {
-    match expr {
-        Expr::Var(name) | Expr::ArrayVar(name) | Expr::HashVar(name) => {
-            name.starts_with('.') || name.starts_with('!')
-        }
-        Expr::MethodCall { target, args, .. } | Expr::HyperMethodCall { target, args, .. } => {
-            expr_uses_attr_twigil(target) || args.iter().any(expr_uses_attr_twigil)
-        }
-        Expr::HyperMethodCallDynamic {
-            target,
-            name_expr,
-            args,
-            ..
-        } => {
-            expr_uses_attr_twigil(target)
-                || expr_uses_attr_twigil(name_expr)
-                || args.iter().any(expr_uses_attr_twigil)
-        }
-        Expr::Call { args, .. }
-        | Expr::ArrayLiteral(args)
-        | Expr::BracketArray(args, _)
-        | Expr::CaptureLiteral(args)
-        | Expr::StringInterpolation(args) => args.iter().any(expr_uses_attr_twigil),
-        Expr::Unary { expr, .. } | Expr::PostfixOp { expr, .. } | Expr::Reduction { expr, .. } => {
-            expr_uses_attr_twigil(expr)
-        }
-        Expr::Binary { left, right, .. }
-        | Expr::MetaOp { left, right, .. }
-        | Expr::HyperOp { left, right, .. } => {
-            expr_uses_attr_twigil(left) || expr_uses_attr_twigil(right)
-        }
-        Expr::Ternary {
-            cond,
-            then_expr,
-            else_expr,
-        } => {
-            expr_uses_attr_twigil(cond)
-                || expr_uses_attr_twigil(then_expr)
-                || expr_uses_attr_twigil(else_expr)
-        }
-        Expr::Index { target, index, .. } => {
-            expr_uses_attr_twigil(target) || expr_uses_attr_twigil(index)
-        }
-        Expr::IndexAssign {
-            target,
-            index,
-            value,
-            ..
-        } => {
-            expr_uses_attr_twigil(target)
-                || expr_uses_attr_twigil(index)
-                || expr_uses_attr_twigil(value)
-        }
-        Expr::AssignExpr { expr, .. } => expr_uses_attr_twigil(expr),
-        Expr::DoBlock { body, .. }
-        | Expr::Block(body)
-        | Expr::Gather(body)
-        | Expr::AnonSub { body, .. }
-        | Expr::AnonSubParams { body, .. }
-        | Expr::Lambda { body, .. } => body.iter().any(stmt_uses_attr_twigil),
-        Expr::Try { body, catch } => {
-            body.iter().any(stmt_uses_attr_twigil)
-                || catch
-                    .as_ref()
-                    .is_some_and(|body| body.iter().any(stmt_uses_attr_twigil))
-        }
-        Expr::DoStmt(stmt) => stmt_uses_attr_twigil(stmt),
-        Expr::CallOn { target, args } => {
-            expr_uses_attr_twigil(target) || args.iter().any(expr_uses_attr_twigil)
-        }
-        Expr::InfixFunc { left, right, .. } => {
-            expr_uses_attr_twigil(left) || right.iter().any(expr_uses_attr_twigil)
-        }
-        Expr::Exists { target, arg, .. } => {
-            expr_uses_attr_twigil(target)
-                || arg
-                    .as_ref()
-                    .is_some_and(|arg_expr| expr_uses_attr_twigil(arg_expr))
-        }
-        Expr::ZenSlice(inner) => expr_uses_attr_twigil(inner),
-        _ => false,
-    }
-}
-
-fn stmt_uses_attr_twigil(stmt: &Stmt) -> bool {
-    match stmt {
-        Stmt::Expr(expr)
-        | Stmt::Return(expr)
-        | Stmt::Take(expr, _)
-        | Stmt::Die(expr)
-        | Stmt::Fail(expr) => expr_uses_attr_twigil(expr),
-        Stmt::VarDecl {
-            expr,
-            where_constraint,
-            ..
-        } => {
-            expr_uses_attr_twigil(expr)
-                || where_constraint
-                    .as_ref()
-                    .is_some_and(|wc| expr_uses_attr_twigil(wc))
-        }
-        Stmt::Assign { expr, .. } => expr_uses_attr_twigil(expr),
-        Stmt::Block(body)
-        | Stmt::SyntheticBlock(body)
-        | Stmt::Package { body, .. }
-        | Stmt::Catch(body)
-        | Stmt::Control(body) => body.iter().any(stmt_uses_attr_twigil),
-        Stmt::SubDecl { body, .. }
-        | Stmt::MethodDecl { body, .. }
-        | Stmt::TokenDecl { body, .. }
-        | Stmt::RuleDecl { body, .. } => body.iter().any(stmt_uses_attr_twigil),
-        Stmt::Label { stmt, .. } => stmt_uses_attr_twigil(stmt),
-        _ => false,
-    }
-}
-
-fn no_self_error() -> PError {
-    let msg = "X::Syntax::NoSelf".to_string();
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    let ex = Value::make_instance(Symbol::intern("X::Syntax::NoSelf"), attrs);
-    PError::fatal_with_exception(msg, Box::new(ex))
-}
-
-/// `X::Syntax::Regex::NullRegex` — an empty `token`/`regex`/`rule` body (e.g.
-/// `regex foo { }`) is a null regex, rejected by Raku at parse time.
-fn null_regex_error() -> PError {
-    let msg = "Null regex not allowed".to_string();
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    let ex = Value::make_instance(Symbol::intern("X::Syntax::Regex::NullRegex"), attrs);
-    PError::fatal_with_exception(msg, Box::new(ex))
-}
-
-fn expr_is_bare_ident(expr: &Expr, ident: &str) -> bool {
-    matches!(expr, Expr::Var(name) | Expr::BareWord(name) if name == ident)
-}
-
-fn stmt_is_also_is_rw(stmt: &Stmt) -> bool {
-    match stmt {
-        Stmt::Expr(Expr::InfixFunc {
-            name, left, right, ..
-        }) => {
-            name == "is"
-                && expr_is_bare_ident(left, "also")
-                && right.len() == 1
-                && expr_is_bare_ident(&right[0], "rw")
-        }
-        Stmt::Expr(Expr::Binary { left, op, right }) => {
-            matches!(op, TokenKind::Ident(name) if name == "is")
-                && expr_is_bare_ident(left, "also")
-                && expr_is_bare_ident(right, "rw")
-        }
-        _ => false,
-    }
-}
-
-/// Extract the parent class name from `also is <ClassName>` statements
-/// (where ClassName is not `rw`).
-fn stmt_also_is_parent(stmt: &Stmt) -> Option<String> {
-    match stmt {
-        Stmt::Expr(Expr::InfixFunc {
-            name, left, right, ..
-        }) if name == "is" && expr_is_bare_ident(left, "also") && right.len() == 1 => {
-            if let Expr::BareWord(parent) = &right[0]
-                && parent != "rw"
-            {
-                return Some(parent.clone());
-            }
-            None
-        }
-        Stmt::Expr(Expr::Binary { left, op, right }) => {
-            if matches!(op, TokenKind::Ident(name) if name == "is")
-                && expr_is_bare_ident(left, "also")
-                && let Expr::BareWord(parent) = right.as_ref()
-                && parent != "rw"
-            {
-                return Some(parent.clone());
-            }
-            None
-        }
-        _ => None,
-    }
-}
-
-fn reject_no_self_in_subs(body: &[Stmt]) -> Result<(), PError> {
-    for stmt in body {
-        if let Stmt::SubDecl { body: sub_body, .. } = stmt
-            && sub_body.iter().any(stmt_uses_attr_twigil)
-        {
-            return Err(no_self_error());
-        }
-    }
-    Ok(())
-}
-
-/// Reject attribute-twigil references (`$!a`, `$.a`, ...) inside a `where`
-/// constraint on an attribute declaration. The `where` clause is evaluated as a
-/// thunk that has no `self`, so such a reference is an X::Syntax::NoSelf error.
-/// Note: this only applies to the `where` constraint, not to the `= default`
-/// initializer (where `$!a` is allowed, since defaults run with `self`).
-fn reject_no_self_in_attr_where(body: &[Stmt]) -> Result<(), PError> {
-    for stmt in body {
-        if let Stmt::HasDecl {
-            where_constraint: Some(wc),
-            ..
-        } = stmt
-            && expr_uses_attr_twigil(wc)
-        {
-            return Err(no_self_error());
-        }
-    }
-    Ok(())
-}
-
-/// Collect no-twigil attribute names from HasDecl statements in a class body.
-fn collect_no_twigil_attr_names(body: &[Stmt]) -> Vec<String> {
-    let mut names = Vec::new();
-    for stmt in body {
-        if let Stmt::HasDecl {
-            name,
-            is_alias: true,
-            ..
-        } = stmt
-        {
-            names.push(name.to_string());
-        }
-    }
-    names
-}
-
-/// Check if an expression references a variable matching one of the given names.
-fn expr_uses_var_name(expr: &Expr, names: &[String]) -> bool {
-    match expr {
-        Expr::Var(name) => names.iter().any(|n| n == name),
-        Expr::MethodCall { target, args, .. } | Expr::HyperMethodCall { target, args, .. } => {
-            expr_uses_var_name(target, names) || args.iter().any(|a| expr_uses_var_name(a, names))
-        }
-        Expr::Call { args, .. }
-        | Expr::ArrayLiteral(args)
-        | Expr::BracketArray(args, _)
-        | Expr::CaptureLiteral(args)
-        | Expr::StringInterpolation(args) => args.iter().any(|a| expr_uses_var_name(a, names)),
-        Expr::Unary { expr, .. } | Expr::PostfixOp { expr, .. } | Expr::Reduction { expr, .. } => {
-            expr_uses_var_name(expr, names)
-        }
-        Expr::Binary { left, right, .. } => {
-            expr_uses_var_name(left, names) || expr_uses_var_name(right, names)
-        }
-        Expr::InfixFunc { left, right, .. } => {
-            expr_uses_var_name(left, names) || right.iter().any(|a| expr_uses_var_name(a, names))
-        }
-        Expr::Ternary {
-            cond,
-            then_expr,
-            else_expr,
-            ..
-        } => {
-            expr_uses_var_name(cond, names)
-                || expr_uses_var_name(then_expr, names)
-                || expr_uses_var_name(else_expr, names)
-        }
-        Expr::AssignExpr { expr, .. } => expr_uses_var_name(expr, names),
-        _ => false,
-    }
-}
-
-/// Check if a statement at class body level uses a no-twigil attribute variable.
-fn stmt_uses_var_name_at_body_level(stmt: &Stmt, names: &[String]) -> bool {
-    match stmt {
-        // Skip method/sub/token/rule declarations — they have `self`
-        Stmt::MethodDecl { .. }
-        | Stmt::SubDecl { .. }
-        | Stmt::TokenDecl { .. }
-        | Stmt::RuleDecl { .. } => false,
-        // Skip HasDecl — declaring the attr is fine
-        Stmt::HasDecl { .. } => false,
-        Stmt::Expr(expr) | Stmt::Return(expr) | Stmt::Fail(expr) => expr_uses_var_name(expr, names),
-        Stmt::Say(args) | Stmt::Print(args) | Stmt::Note(args) | Stmt::Put(args) => {
-            args.iter().any(|a| expr_uses_var_name(a, names))
-        }
-        _ => false,
-    }
-}
-
-/// Reject no-twigil attribute variable usage at class body level (outside methods).
-fn reject_no_twigil_attr_at_body_level(body: &[Stmt]) -> Result<(), PError> {
-    let names = collect_no_twigil_attr_names(body);
-    if names.is_empty() {
-        return Ok(());
-    }
-    for stmt in body {
-        if stmt_uses_var_name_at_body_level(stmt, &names) {
-            return Err(no_self_error());
-        }
-    }
-    Ok(())
-}
-
-fn consume_raw_braced_body(input: &str) -> PResult<'_, Vec<Stmt>> {
-    if !input.starts_with('{') {
-        return Err(PError::expected("raw braced body"));
-    }
-    let mut depth = 0u32;
-    let mut i = 0usize;
-    while i < input.len() {
-        let ch = input[i..]
-            .chars()
-            .next()
-            .ok_or_else(|| PError::expected("closing '}'"))?;
-        let len = ch.len_utf8();
-        match ch {
-            '{' => depth += 1,
-            '}' => {
-                depth = depth.saturating_sub(1);
-                if depth == 0 {
-                    let rest = &input[i + len..];
-                    return Ok((rest, Vec::new()));
-                }
-            }
-            '\\' => {
-                i += len;
-                if i < input.len() {
-                    let next_len = input[i..].chars().next().map(|c| c.len_utf8()).unwrap_or(0);
-                    i += next_len;
-                    continue;
-                }
-            }
-            '\'' | '"' => {
-                let quote = ch;
-                i += len;
-                while i < input.len() {
-                    let c = input[i..]
-                        .chars()
-                        .next()
-                        .ok_or_else(|| PError::expected("string close"))?;
-                    let c_len = c.len_utf8();
-                    if c == '\\' {
-                        i += c_len;
-                        if i < input.len() {
-                            let n_len =
-                                input[i..].chars().next().map(|n| n.len_utf8()).unwrap_or(0);
-                            i += n_len;
-                            continue;
-                        }
-                    }
-                    i += c_len;
-                    if c == quote {
-                        break;
-                    }
-                }
-                continue;
-            }
-            _ => {}
-        }
-        i += len;
-    }
-    Err(PError::expected("closing '}'"))
-}
-
-fn parse_token_like_name(input: &str) -> PResult<'_, String> {
-    let (mut rest, mut name) = ident(input)?;
-    loop {
-        // `::`-qualified name segments (e.g. `does DBDish::ErrorHandling`).
-        // Handle these before the single-`:` adverb logic so a qualified role
-        // name parses as one token rather than failing on the second segment.
-        if rest.starts_with("::") && !rest[2..].starts_with('(') {
-            if let Ok((r, seg)) = ident(&rest[2..]) {
-                name.push_str("::");
-                name.push_str(&seg);
-                rest = r;
-                continue;
-            }
-            break;
-        }
-        if !rest.starts_with(':') {
-            break;
-        }
-        let r = &rest[1..];
-        // `:sym<value>` has an identifier part; `:<value>` (shorthand for
-        // `:sym<value>`) has the colon immediately followed by `<...>` with no
-        // part name. Allow an empty part in that case.
-        let (r, part) = if r.starts_with('<') || r.starts_with("<<") || r.starts_with('\u{ab}') {
-            (r, "")
-        } else {
-            take_while1(r, |c: char| c.is_alphanumeric() || c == '_' || c == '-')?
-        };
-        name.push(':');
-        name.push_str(part);
-        let mut r2 = r;
-        if r2.starts_with("<<") {
-            // <<...>> delimiter (ASCII double-angle quotes)
-            let after_open = &r2[2..];
-            if let Some(end) = after_open.find(">>") {
-                // Store as «...» internally for consistency
-                name.push('\u{ab}');
-                name.push_str(&after_open[..end]);
-                name.push('\u{bb}');
-                r2 = &after_open[end + 2..];
-            }
-        } else if r2.starts_with('<')
-            && let Some(end) = r2.find('>')
-        {
-            name.push_str(&r2[..=end]);
-            r2 = &r2[end + 1..];
-        } else if r2.starts_with('\u{ab}') {
-            // «» (French quotes) — keep as «» internally to avoid
-            // ambiguity when the value contains '>'
-            let after_open = &r2['\u{ab}'.len_utf8()..];
-            if let Some(end) = after_open.find('\u{bb}') {
-                name.push('\u{ab}');
-                name.push_str(&after_open[..end]);
-                name.push('\u{bb}');
-                r2 = &after_open[end + '\u{bb}'.len_utf8()..];
-            }
-        }
-        rest = r2;
-    }
-    Ok((rest, name))
-}
-
-fn parse_raw_braced_regex_body(input: &str) -> PResult<'_, String> {
-    let after_open = input
-        .strip_prefix('{')
-        .ok_or_else(|| PError::expected("regex body"))?;
-    if let Some((body, rest)) = scan_to_delim(after_open, '{', '}', true) {
-        return Ok((rest, body.trim().to_string()));
-    }
-    Err(PError::expected("regex closing delimiter"))
-}
-
-fn inject_implicit_rule_ws(pattern: &str) -> String {
-    fn should_insert(prev: char, next: char) -> bool {
-        !matches!(
-            (prev, next),
-            ('|', _)
-                | (_, '|')
-                | ('(', _)
-                | (_, ')')
-                | ('[', _)
-                | (_, ']')
-                | ('{', _)
-                | (_, '}')
-                | ('^', _)
-                | (_, '$')
-                | ('<', _)
-                | (_, '>')
-                | (_, '*')
-                | (_, '+')
-                | (_, '?')
-                | (_, '%')
-                | ('%', _)
-        )
-    }
-
-    let chars: Vec<char> = pattern.chars().collect();
-    let mut out = String::new();
-    let mut i = 0usize;
-    let mut in_single = false;
-    let mut in_double = false;
-    let mut escaped = false;
-    let mut brace_depth = 0usize;
-    while i < chars.len() {
-        let c = chars[i];
-        if escaped {
-            out.push(c);
-            escaped = false;
-            i += 1;
-            continue;
-        }
-        if c == '\\' {
-            out.push(c);
-            escaped = true;
-            i += 1;
-            continue;
-        }
-        if c == '\'' && !in_double {
-            in_single = !in_single;
-            out.push(c);
-            i += 1;
-            continue;
-        }
-        if c == '"' && !in_single {
-            in_double = !in_double;
-            out.push(c);
-            i += 1;
-            continue;
-        }
-        // Track brace depth to skip ws injection inside code blocks { ... }
-        if !in_single && !in_double {
-            if c == '{' {
-                brace_depth += 1;
-                out.push(c);
-                i += 1;
-                continue;
-            }
-            if c == '}' && brace_depth > 0 {
-                brace_depth -= 1;
-                out.push(c);
-                i += 1;
-                continue;
-            }
-        }
-        // Inside a code block — pass through without ws injection
-        if brace_depth > 0 {
-            out.push(c);
-            i += 1;
-            continue;
-        }
-        if !in_single && !in_double && c.is_whitespace() {
-            let mut j = i;
-            while j < chars.len() && chars[j].is_whitespace() {
-                j += 1;
-            }
-            let prev = out.chars().rev().find(|ch| !ch.is_whitespace());
-            let next = chars[j..].iter().copied().find(|ch| !ch.is_whitespace());
-            if let (Some(p), Some(n)) = (prev, next) {
-                if p == '^' {
-                    if !out.ends_with(' ') && !out.is_empty() {
-                        out.push(' ');
-                    }
-                    out.push_str("<.ws>?");
-                    out.push(' ');
-                } else if should_insert(p, n) {
-                    if !out.ends_with(' ') && !out.is_empty() {
-                        out.push(' ');
-                    }
-                    out.push_str("<.ws>");
-                    out.push(' ');
-                } else if !out.ends_with(' ') && !out.is_empty() {
-                    out.push(' ');
-                }
-            }
-            i = j;
-            continue;
-        }
-        out.push(c);
-        i += 1;
-    }
-    out.trim().to_string()
-}
-
-/// In a `rule`, the `%` separator quantifier should allow optional whitespace
-/// around the separator. This function finds `% SEPARATOR` patterns and wraps
-/// the separator with `[ <.ws>? SEPARATOR <.ws>? ]`.
-fn inject_separator_ws(pattern: &str) -> String {
-    let chars: Vec<char> = pattern.chars().collect();
-    let mut out = String::new();
-    let mut i = 0usize;
-    let mut escaped = false;
-    let mut in_single = false;
-    let mut in_double = false;
-    while i < chars.len() {
-        let c = chars[i];
-        if escaped {
-            out.push(c);
-            escaped = false;
-            i += 1;
-            continue;
-        }
-        if c == '\\' {
-            out.push(c);
-            escaped = true;
-            i += 1;
-            continue;
-        }
-        if c == '\'' && !in_double {
-            in_single = !in_single;
-            out.push(c);
-            i += 1;
-            continue;
-        }
-        if c == '"' && !in_single {
-            in_double = !in_double;
-            out.push(c);
-            i += 1;
-            continue;
-        }
-        if in_single || in_double {
-            out.push(c);
-            i += 1;
-            continue;
-        }
-        // Look for `%` (or `%%`) followed by a separator expression
-        if c == '%' {
-            out.push('%');
-            i += 1;
-            // `%%` (trailing-separator-allowed) is a single operator: emit its
-            // second `%` too, otherwise the second `%` is mistaken for the
-            // separator atom and the operator is destroyed (e.g.
-            // `\d+ %% "+"` would become `\d+ % [<.ws>? % <.ws>?] "+"`).
-            if i < chars.len() && chars[i] == '%' {
-                out.push('%');
-                i += 1;
-            }
-            // Skip whitespace after % / %%
-            while i < chars.len() && chars[i].is_whitespace() {
-                out.push(chars[i]);
-                i += 1;
-            }
-            if i >= chars.len() {
-                continue;
-            }
-            // Extract the separator expression
-            let sep_start = i;
-            if chars[i] == '[' {
-                // Bracketed separator: % [ ... ]
-                // Scan to matching ]
-                let mut depth = 1usize;
-                i += 1;
-                while i < chars.len() && depth > 0 {
-                    if chars[i] == '[' {
-                        depth += 1;
-                    } else if chars[i] == ']' {
-                        depth -= 1;
-                    } else if chars[i] == '\\' {
-                        i += 1; // skip escaped char
-                    }
-                    i += 1;
-                }
-                let sep: String = chars[sep_start..i].iter().collect();
-                out.push_str(&format!("[ <.ws>? {} <.ws>? ]", sep.trim()));
-            } else {
-                // Non-bracketed separator: % \, or % ","
-                // Collect the separator atom (could be \X, 'str', "str", or a single char)
-                let atom_start = i;
-                if chars[i] == '\'' {
-                    // Single-quoted string
-                    i += 1;
-                    while i < chars.len() && chars[i] != '\'' {
-                        if chars[i] == '\\' {
-                            i += 1;
-                        }
-                        i += 1;
-                    }
-                    if i < chars.len() {
-                        i += 1;
-                    } // skip closing '
-                } else if chars[i] == '"' {
-                    // Double-quoted string
-                    i += 1;
-                    while i < chars.len() && chars[i] != '"' {
-                        if chars[i] == '\\' {
-                            i += 1;
-                        }
-                        i += 1;
-                    }
-                    if i < chars.len() {
-                        i += 1;
-                    } // skip closing "
-                } else if chars[i] == '\\' {
-                    // Backslash-escaped char
-                    i += 2;
-                } else if chars[i] == '<' {
-                    // Angle-bracket assertion
-                    let mut depth = 1usize;
-                    i += 1;
-                    while i < chars.len() && depth > 0 {
-                        if chars[i] == '<' {
-                            depth += 1;
-                        } else if chars[i] == '>' {
-                            depth -= 1;
-                        }
-                        i += 1;
-                    }
-                } else {
-                    // Single character separator
-                    i += 1;
-                }
-                let sep: String = chars[atom_start..i].iter().collect();
-                out.push_str(&format!("[ <.ws>? {} <.ws>? ]", sep.trim()));
-            }
-            continue;
-        }
-        out.push(c);
-        i += 1;
-    }
-    out
-}
-
-fn normalize_token_pattern(pattern: &str) -> String {
-    let trimmed = pattern.trim();
-    if trimmed.len() >= 2 && trimmed.starts_with('/') && trimmed.ends_with('/') {
-        trimmed[1..trimmed.len() - 1].to_string()
-    } else {
-        trimmed.to_string()
-    }
-}
-
-/// Parse `class` declaration.
-pub(crate) fn class_decl(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("class", input).ok_or_else(|| PError::expected("class declaration"))?;
-    let (rest, _) = ws1(rest)?;
-    class_decl_body(rest, false)
-}
-
-/// Parse `augment class ClassName { ... }` declaration (monkey-patching).
-/// Also handles `augment role RoleName { ... }` (always illegal — roles are
-/// closed) and the anonymous form `augment class { ... }` (X::Anon::Augment).
-pub(crate) fn augment_class_decl(input: &str) -> PResult<'_, Stmt> {
-    let rest =
-        keyword("augment", input).ok_or_else(|| PError::expected("augment class declaration"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, is_role) = if let Some(r) = keyword("class", rest) {
-        (r, false)
-    } else if let Some(r) = keyword("role", rest) {
-        (r, true)
-    } else {
-        return Err(PError::expected("'class' or 'role' after 'augment'"));
-    };
-    let (rest, _) = ws1(rest)?;
-    // Anonymous augment (`augment class { }` / `augment role { }`) — no name
-    // follows the package kind. Raku rejects this with X::Anon::Augment.
-    let package_kind = if is_role { "role" } else { "class" };
-    if rest.starts_with('{') {
-        let msg = format!("Cannot augment anonymous {}", package_kind);
-        let mut attrs = std::collections::HashMap::new();
-        attrs.insert("message".to_string(), Value::str(msg.clone()));
-        attrs.insert(
-            "package-kind".to_string(),
-            Value::str(package_kind.to_string()),
-        );
-        let ex = Value::make_instance(Symbol::intern("X::Anon::Augment"), attrs);
-        return Err(PError::fatal_with_exception(msg, Box::new(ex)));
-    }
-    let (rest, name) = qualified_ident(rest)?;
-    // Check for type adverbs (:D, :U, :auth, :ver, :api) which are not allowed on augment
-    if rest.starts_with(':') && !rest.starts_with("::") {
-        let adverb_rest = &rest[1..];
-        // Check for known type adverbs
-        let is_adverb = adverb_rest.starts_with('D')
-            || adverb_rest.starts_with('U')
-            || adverb_rest.starts_with("auth")
-            || adverb_rest.starts_with("ver")
-            || adverb_rest.starts_with("api");
-        if is_adverb {
-            let msg = "Cannot put adverbs on a typename when augmenting".to_string();
-            let mut attrs = std::collections::HashMap::new();
-            attrs.insert("message".to_string(), Value::str(msg.clone()));
-            let ex = Value::make_instance(Symbol::intern("X::Syntax::Augment::Adverb"), attrs);
-            return Err(PError::fatal_with_exception(msg, Box::new(ex)));
-        }
-    }
-    let (rest, _) = ws(rest)?;
-    let (rest, body) = block(rest)?;
-    Ok((
-        rest,
-        Stmt::AugmentClass {
-            name: Symbol::intern(&name),
-            body,
-            is_role,
-        },
-    ))
-}
-
-/// Parse `anon class Name { ... }` declaration.
-/// The class is created but not installed in the lexical scope.
-/// Emits a ClassDecl (registered under the real name) followed by a call to
-/// `__MUTSU_UNREGISTER_CLASS__` to remove the name from the scope.
-pub(crate) fn anon_class_decl(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("anon", input).ok_or_else(|| PError::expected("anon class declaration"))?;
-    let (rest, _) = ws1(rest)?;
-    let rest = keyword("class", rest).ok_or_else(|| PError::expected("class after anon"))?;
-    let (rest, _) = ws1(rest)?;
-    // Parse the class name
-    let (rest, user_name) = qualified_ident(rest)?;
-    let (rest, _) = ws(rest)?;
-    // Parse parent classes
-    let mut parents = Vec::new();
-    let mut anon_repr: Option<String> = None;
-    let mut r = rest;
-    while let Some(r2) = keyword("is", r) {
-        let (r2, _) = ws1(r2)?;
-        let (r2, parent) = qualified_ident(r2)?;
-        if parent == "repr" {
-            if let Some(inner) = r2.strip_prefix('(') {
-                let end = inner.find(')').unwrap_or(inner.len());
-                let repr_val = inner[..end].trim().trim_matches('\'').trim_matches('"');
-                anon_repr = Some(repr_val.to_string());
-            }
-            let r2 = skip_balanced_parens(r2);
-            let (r2, _) = ws(r2)?;
-            r = r2;
-            continue;
-        }
-        parents.push(parent);
-        let (r2, _) = ws(r2)?;
-        r = r2;
-    }
-    let (rest, body) = block(r)?;
-    reject_no_self_in_subs(&body)?;
-    reject_no_self_in_attr_where(&body)?;
-    reject_no_twigil_attr_at_body_level(&body)?;
-    let class_decl = Stmt::ClassDecl {
-        name: Symbol::intern(&user_name),
-        name_expr: None,
-        parents,
-        class_is_rw: false,
-        is_hidden: false,
-        is_lexical: false,
-        hidden_parents: Vec::new(),
-        does_parents: Vec::new(),
-        repr: anon_repr,
-        body,
-        language_version: super::simple::current_language_version(),
-        custom_traits: Vec::new(),
-        is_unit: false,
-    };
-    // Emit the class registration followed by unregistering the name from the scope
-    let unregister = Stmt::Expr(Expr::Call {
-        name: Symbol::intern("__MUTSU_UNREGISTER_CLASS__"),
-        args: vec![Expr::Literal(Value::str(user_name))],
-    });
-    Ok((rest, Stmt::Block(vec![class_decl, unregister])))
-}
-
-/// Parse the body of a class declaration (after `class` keyword and whitespace).
-pub(super) fn class_decl_body(input: &str, is_lexical: bool) -> PResult<'_, Stmt> {
-    let (rest, name, name_expr) = if let Some(after_colons) = input.strip_prefix("::") {
-        // Check if this is `::Ident` (forward stub) or `::(expr)` (indirect name)
-        if after_colons.starts_with('(') {
-            let (rest, (name, expr)) = parse_indirect_decl_name(input)?;
-            (rest, name, Some(expr))
-        } else {
-            // `class ::F { ... }` — forward declaration stub, name is just the identifier
-            let (rest, name) = qualified_ident(after_colons)?;
-            (rest, name, None)
-        }
-    } else {
-        let (rest, name) = qualified_ident(input)?;
-        (rest, name, None)
-    };
-    check_pseudo_package_in_decl(&name)?;
-    let (rest, traits) = parse_declarator_traits(rest)?;
-    let (rest, _) = ws(rest)?;
-
-    // Parent clauses in any order: `is Parent`, `does Role[...]`, `hides Parent`.
-    let mut is_hidden = false;
-    let mut class_is_rw = false;
-    let mut hidden_parents = Vec::new();
-    let mut parents = Vec::new();
-    let mut does_parents = Vec::new();
-    let mut is_repr: Option<String> = None;
-    let mut custom_traits: Vec<(String, Option<Expr>)> = Vec::new();
-    let mut r = rest;
-    loop {
-        if let Some(r2) = keyword("is", r) {
-            let (r2, _) = ws1(r2)?;
-            // Handle `is ::Foo` (indirect name lookup) — treat the
-            // `::Ident` as a parent name so that validation later
-            // produces X::Inheritance::UnknownParent.
-            let (r2, parent) = if let Some(stripped) = r2.strip_prefix("::") {
-                let (r3, ident_part) = qualified_ident(stripped)?;
-                (r3, format!("::{}", ident_part))
-            } else {
-                qualified_ident(r2)?
-            };
-            if parent == "hidden" {
-                is_hidden = true;
-            } else if parent == "rw" {
-                class_is_rw = true;
-            } else if parent == "repr" {
-                // Extract repr value from `is repr('CUnion')` etc.
-                if let Some(inner) = r2.strip_prefix('(') {
-                    // Find the content between parens, stripping quotes
-                    let end = inner.find(')').unwrap_or(inner.len());
-                    let repr_val = inner[..end].trim().trim_matches('\'').trim_matches('"');
-                    is_repr = Some(repr_val.to_string());
-                }
-                let r2 = skip_balanced_parens(r2);
-                let (r2, _) = ws(r2)?;
-                r = r2;
-                continue;
-            } else if parent == "DEPRECATED" {
-                // `is DEPRECATED` on a class — skip optional parenthesized arg
-                let r2 = skip_balanced_parens(r2);
-                let (r2, _) = ws(r2)?;
-                r = r2;
-                continue;
-            } else if parent.starts_with(|c: char| c.is_ascii_uppercase())
-                || parent.starts_with("::")
-            {
-                let (r2, bracket_suffix) = parse_optional_bracket_suffix(r2)?;
-                parents.push(format!("{}{}", parent, bracket_suffix));
-                r = r2;
-                let (r2, _) = ws(r)?;
-                r = r2;
-                continue;
-            }
-            // Unrecognized `is` trait with lowercase name — check for custom
-            // trait_mod:<is> dispatch with optional parenthesized argument, or
-            // treat as a potential parent class.
-            if !matches!(
-                parent.as_str(),
-                "rw" | "hidden"
-                    | "export"
-                    | "DEPRECATED"
-                    | "default"
-                    | "nodal"
-                    | "raw"
-                    | "required"
-                    | "built"
-                    | "cached"
-                    | "repr"
-                    | "open"
-            ) {
-                // Check for parenthesized argument: `is trait-name('arg')`
-                // This indicates a custom trait_mod:<is> call.
-                if let Some(inner_start) = r2.strip_prefix('(') {
-                    if let Ok((after_expr, expr)) = expression(inner_start) {
-                        let after_expr = after_expr.trim_start();
-                        if let Some(after_paren) = after_expr.strip_prefix(')') {
-                            custom_traits.push((parent.clone(), Some(expr)));
-                            let (r3, _) = ws(after_paren)?;
-                            r = r3;
-                            continue;
-                        }
-                    }
-                    // Fallback: skip balanced parens
-                    let r3 = skip_balanced_parens(r2);
-                    custom_traits.push((parent.clone(), None));
-                    let (r3, _) = ws(r3)?;
-                    r = r3;
-                    continue;
-                }
-                let (r2, bracket_suffix) = parse_optional_bracket_suffix(r2)?;
-                parents.push(format!("{}{}", parent, bracket_suffix));
-                r = r2;
-                let (r2, _) = ws(r)?;
-                r = r2;
-                continue;
-            }
-            let (r2, _) = ws(r2)?;
-            r = r2;
-            continue;
-        }
-        if let Some(r2) = keyword("does", r) {
-            let (r2, _) = ws1(r2)?;
-            let (r2, role_name) = qualified_ident(r2)?;
-            let (r2, _) = ws(r2)?;
-            let (r2, bracket_suffix) = parse_optional_bracket_suffix(r2)?;
-            let full_name = format!("{}{}", role_name, bracket_suffix);
-            parents.push(full_name.clone());
-            does_parents.push(full_name);
-            let (r2, _) = ws(r2)?;
-            r = r2;
-            continue;
-        }
-        if let Some(r2) = keyword("hides", r) {
-            let (r2, _) = ws1(r2)?;
-            let (r2, parent) = qualified_ident(r2)?;
-            let (r2, _) = ws(r2)?;
-            let (r2, bracket_suffix) = parse_optional_bracket_suffix(r2)?;
-            let hidden_parent = format!("{}{}", parent, bracket_suffix);
-            parents.push(hidden_parent.clone());
-            hidden_parents.push(hidden_parent);
-            let (r2, _) = ws(r2)?;
-            r = r2;
-            continue;
-        }
-        break;
-    }
-
-    let (rest, mut body) = block(r)?;
-    reject_no_self_in_subs(&body)?;
-    reject_no_self_in_attr_where(&body)?;
-    reject_no_twigil_attr_at_body_level(&body)?;
-    body.retain(|stmt| {
-        if stmt_is_also_is_rw(stmt) {
-            class_is_rw = true;
-            false
-        } else if let Some(parent_name) = stmt_also_is_parent(stmt) {
-            parents.push(parent_name);
-            false
-        } else {
-            true
-        }
-    });
-    // Extract repr from `is repr(...)` or from declarator traits
-    let repr = is_repr.or_else(|| {
-        traits.iter().find_map(|(k, v)| {
-            if k == "repr" {
-                if let Value::Str(s) = v {
-                    Some(s.to_string())
-                } else {
-                    None
-                }
-            } else {
-                None
-            }
-        })
-    });
-    // Register the class name so the parser can disambiguate identifiers
-    // from regex/substitution operators (e.g. `S` vs `S///`).
-    super::simple::register_user_type(&name);
-
-    // Validate declarator traits: only 'ver', 'auth', and 'api' are allowed on class names
-    for (trait_name, _) in &traits {
-        if trait_name != "ver" && trait_name != "auth" && trait_name != "api" {
-            let msg = format!(
-                "Cannot use adverb {} on a type name (only 'ver', 'auth' and 'api' are understood)",
-                trait_name
-            );
-            let mut attrs = std::collections::HashMap::new();
-            attrs.insert("message".to_string(), Value::str(msg.clone()));
-            let ex = Value::make_instance(Symbol::intern("X::Syntax::Type::Adverb"), attrs);
-            return Err(PError::fatal_with_exception(msg, Box::new(ex)));
-        }
-    }
-    let class_stmt = Stmt::ClassDecl {
-        name: Symbol::intern(&name),
-        name_expr,
-        parents,
-        class_is_rw,
-        is_hidden,
-        is_lexical,
-        hidden_parents,
-        does_parents,
-        repr,
-        body,
-        language_version: super::simple::current_language_version(),
-        custom_traits,
-        is_unit: false,
-    };
-    let mut stmts = Vec::new();
-    for (trait_name, trait_value) in traits {
-        if trait_name == "ver" || trait_name == "auth" {
-            stmts.push(meta_setter_stmt(&name, &trait_name, trait_value));
-        }
-    }
-    if stmts.is_empty() {
-        return Ok((rest, class_stmt));
-    }
-    stmts.push(class_stmt);
-    Ok((rest, Stmt::Block(stmts)))
-}
-
-/// Parse `also is <trait>;` statement.
-///
-/// This is primarily used inside class bodies (e.g. `also is rw;`) and is
-/// represented as a small infix expression statement for later lowering.
-pub(super) fn also_trait_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("also", input).ok_or_else(|| PError::expected("also trait statement"))?;
-    let (rest, _) = ws1(rest)?;
-    // Handle `also does RoleName;`
-    if let Some(r) = keyword("does", rest) {
-        let (r, _) = ws1(r)?;
-        let (r, name) = parse_token_like_name(r)?;
-        let (r, _) = ws(r)?;
-        let (r, _) = opt_char(r, ';');
-        return Ok((
-            r,
-            Stmt::DoesDecl {
-                name: Symbol::intern(&name),
-            },
-        ));
-    }
-    let rest = keyword("is", rest).ok_or_else(|| PError::expected("is or does after also"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, trait_name) = ident(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (rest, _) = opt_char(rest, ';');
-    Ok((
-        rest,
-        Stmt::Expr(Expr::InfixFunc {
-            name: "is".to_string(),
-            left: Box::new(Expr::BareWord("also".to_string())),
-            right: vec![Expr::BareWord(trait_name)],
-            modifier: None,
-        }),
-    ))
-}
-
-/// Parse optional role type parameters like `[::T]`, `[Str $x]`, or
-/// `[Int $x where { ... }]`.
-/// Returns both full parameter defs and plain names used for substitution.
-fn parse_optional_role_type_params(input: &str) -> PResult<'_, (Vec<String>, Vec<ParamDef>)> {
-    let (r, _) = ws(input)?;
-    if !r.starts_with('[') {
-        return Ok((r, (Vec::new(), Vec::new())));
-    }
-    let mut depth = 0u32;
-    let mut end = 0usize;
-    for (i, ch) in r.char_indices() {
-        if ch == '[' {
-            depth += 1;
-        } else if ch == ']' {
-            depth = depth.saturating_sub(1);
-            if depth == 0 {
-                end = i;
-                break;
-            }
-        }
-    }
-    if end == 0 {
-        return Err(PError::expected("']'"));
-    }
-    let content = &r[1..end];
-    let normalized_content = content
-        .replace("= my role {", "= role {")
-        .replace("= my class {", "= class {")
-        .replace("= my grammar {", "= grammar {");
-    for part in normalized_content.split(',') {
-        let trimmed = part.trim();
-        if let Some(stripped) = trimmed.strip_prefix("::")
-            && let Ok((rest_after_ident, _)) = ident(stripped)
-            && rest_after_ident.starts_with('?')
-        {
-            return Err(PError::fatal("X::Syntax::Malformed".to_string()));
-        }
-    }
-    if normalized_content == content
-        && let Ok((after_params, param_defs)) = parse_param_list(&r[1..])
-        && let Ok((rest, _)) = parse_char(after_params, ']')
-    {
-        for pd in &param_defs {
-            if pd.name == "__type_only__"
-                && let Some(tc) = pd.type_constraint.as_deref()
-                && !tc.starts_with("::")
-                && !role_type_param_constraint_is_known(tc)
-            {
-                let msg = format!(
-                    "X::Parameter::InvalidType: Invalid type '{}' in role parameter list",
-                    tc
-                );
-                let mut attrs = std::collections::HashMap::new();
-                attrs.insert("type".to_string(), Value::str(tc.to_string()));
-                attrs.insert("message".to_string(), Value::str(msg.clone()));
-                let ex = Value::make_instance(Symbol::intern("X::Parameter::InvalidType"), attrs);
-                return Err(PError::fatal_with_exception(msg, Box::new(ex)));
-            }
-        }
-        let params = param_defs
-            .iter()
-            .map(|pd| {
-                if let Some(captured) = pd
-                    .type_constraint
-                    .as_deref()
-                    .and_then(|t| t.strip_prefix("::"))
-                {
-                    captured.to_string()
-                } else {
-                    pd.name.trim_start_matches(['$', '@', '%', '&']).to_string()
-                }
-            })
-            .collect::<Vec<_>>();
-        let (rest, _) = ws(rest)?;
-        return Ok((rest, (params, param_defs)));
-    }
-
-    // Fallback: parse each top-level parameter item independently so defaults
-    // like `::T = 42` or `::T = my role { ... }` still preserve ParamDef data.
-    let mut params = Vec::new();
-    let mut param_defs = Vec::new();
-    for part in split_role_param_parts(&normalized_content) {
-        if part.is_empty() {
-            continue;
-        }
-        if let Some((constraint_part, capture_part)) = part.split_once("::") {
-            let constraint = constraint_part.trim();
-            let capture_part = capture_part.trim();
-            if !constraint.is_empty()
-                && role_type_param_constraint_is_known(constraint)
-                && let Ok((rest_after_name, name)) = ident(capture_part)
-            {
-                let rest_after_name = rest_after_name.trim_start();
-                let default = if let Some(rhs) = rest_after_name.strip_prefix('=') {
-                    let rhs = rhs.trim_start();
-                    if let Ok((rest_after_expr, default_expr)) = expression(rhs) {
-                        if rest_after_expr.trim().is_empty() {
-                            Some(default_expr)
-                        } else {
-                            continue;
-                        }
-                    } else {
-                        continue;
-                    }
-                } else if rest_after_name.is_empty() {
-                    None
-                } else {
-                    continue;
-                };
-                params.push(name.clone());
-                param_defs.push(ParamDef {
-                    name,
-                    default,
-                    multi_invocant: true,
-                    required: false,
-                    named: false,
-                    slurpy: false,
-                    double_slurpy: false,
-                    onearg: false,
-                    sigilless: false,
-                    type_constraint: Some(constraint.to_string()),
-                    literal_value: None,
-                    sub_signature: None,
-                    where_constraint: None,
-                    traits: Vec::new(),
-                    optional_marker: false,
-                    outer_sub_signature: None,
-                    code_signature: None,
-                    is_invocant: false,
-                    shape_constraints: None,
-                });
-                continue;
-            }
-        }
-        if let Some(after_capture) = part.strip_prefix("::")
-            && let Ok((rest_after_name, name)) = ident(after_capture)
-        {
-            let rest_after_name = rest_after_name.trim_start();
-            if let Some(rhs) = rest_after_name.strip_prefix('=') {
-                let rhs = rhs.trim_start();
-                if let Ok((rest_after_expr, default_expr)) = expression(rhs)
-                    && rest_after_expr.trim().is_empty()
-                {
-                    params.push(name.clone());
-                    param_defs.push(ParamDef {
-                        name: format!("__type_capture__{}", name),
-                        default: Some(default_expr),
-                        multi_invocant: true,
-                        required: false,
-                        named: false,
-                        slurpy: false,
-                        double_slurpy: false,
-                        onearg: false,
-                        sigilless: false,
-                        type_constraint: Some(format!("::{}", name)),
-                        literal_value: None,
-                        sub_signature: None,
-                        where_constraint: None,
-                        traits: Vec::new(),
-                        optional_marker: false,
-                        outer_sub_signature: None,
-                        code_signature: None,
-                        is_invocant: false,
-                        shape_constraints: None,
-                    });
-                    continue;
-                }
-            }
-        }
-        if let Ok((rest, pd)) = parse_single_param(part)
-            && rest.trim().is_empty()
-        {
-            let name = if let Some(captured) = pd
-                .type_constraint
-                .as_deref()
-                .and_then(|t| t.strip_prefix("::"))
-            {
-                captured.to_string()
-            } else {
-                pd.name.trim_start_matches(['$', '@', '%', '&']).to_string()
-            };
-            params.push(name);
-            param_defs.push(pd);
-            continue;
-        }
-        if let Some(stripped) = part.strip_prefix("::") {
-            let name_part = stripped.trim();
-            if let Ok((_, name)) = ident(name_part) {
-                params.push(name);
-            }
-        }
-    }
-    let (rest, _) = ws(&r[end + 1..])?;
-    Ok((rest, (params, param_defs)))
-}
-
-/// Skip optional role args like `[Str:D(Numeric)]` in a `does` clause.
-fn skip_optional_role_args(input: &str) -> PResult<'_, ()> {
-    let (mut r, _) = ws(input)?;
-    if !r.starts_with('[') {
-        return Ok((r, ()));
-    }
-    let mut depth = 0u32;
-    while let Some(ch) = r.chars().next() {
-        let len = ch.len_utf8();
-        if ch == '[' {
-            depth += 1;
-        } else if ch == ']' {
-            depth = depth.saturating_sub(1);
-            if depth == 0 {
-                r = &r[len..];
-                break;
-            }
-        }
-        r = &r[len..];
-    }
-    let (r, _) = ws(r)?;
-    Ok((r, ()))
-}
-
-/// Parse `role` declaration.
-pub(super) fn role_decl(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("role", input).ok_or_else(|| PError::expected("role declaration"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, name) = qualified_ident(rest)?;
-    check_pseudo_package_in_decl(&name)?;
-    let (mut rest, (type_params, type_param_defs)) = parse_optional_role_type_params(rest)?;
-    // Parse optional type adverbs (:ver<...>, :auth<...>, :api<...>)
-    let (rest2, traits) = parse_declarator_traits(rest)?;
-    let (rest2, _) = ws(rest2)?;
-    rest = rest2;
-    let mut parent_roles: Vec<String> = Vec::new();
-    let mut is_hidden_role = false;
-    let mut role_is_rw = false;
-    let mut is_export = false;
-    let mut export_tags: Vec<String> = Vec::new();
-    let mut custom_traits: Vec<(String, Option<Expr>)> = Vec::new();
-
-    // Optional parent/trait clauses in any order.
-    loop {
-        if let Some(r) = keyword("does", rest) {
-            let (r, _) = ws1(r)?;
-            let (r, role_name) = qualified_ident(r)?;
-            if role_name == name {
-                return Err(PError::fatal(format!(
-                    "X::InvalidType: role '{}' cannot compose itself",
-                    name
-                )));
-            }
-            let (r, _) = ws(r)?;
-            let (r, bracket_suffix) = parse_optional_bracket_suffix(r)?;
-            let (r, _) = ws(r)?;
-            parent_roles.push(format!("{}{}", role_name, bracket_suffix));
-            rest = r;
-            continue;
-        }
-        if let Some(r) = keyword("is", rest) {
-            let (r, _) = ws1(r)?;
-            let (r, trait_name) = ident(r)?;
-            if trait_name == "hidden" {
-                is_hidden_role = true;
-                let r = skip_balanced_parens(r);
-                let (r, _) = ws(r)?;
-                rest = r;
-            } else if trait_name == "rw" {
-                role_is_rw = true;
-                let r = skip_balanced_parens(r);
-                let (r, _) = ws(r)?;
-                rest = r;
-            } else if trait_name == "export" {
-                is_export = true;
-                if !export_tags.iter().any(|t| t == "DEFAULT") {
-                    export_tags.push("DEFAULT".to_string());
-                }
-                let r = skip_balanced_parens(r);
-                let (r, _) = ws(r)?;
-                rest = r;
-            } else if matches!(
-                trait_name.as_str(),
-                "ok" | "required"
-                    | "readonly"
-                    | "repr"
-                    | "default"
-                    | "raw"
-                    | "built"
-                    | "copy"
-                    | "DEPRECATED"
-                    | "nodal"
-                    | "pure"
-            ) {
-                // Known lowercase trait keywords are skipped
-                let r = skip_balanced_parens(r);
-                let (r, _) = ws(r)?;
-                rest = r;
-            } else {
-                // Unknown trait — check for parenthesized argument for
-                // custom trait_mod:<is> dispatch.
-                if let Some(inner_start) = r.strip_prefix('(') {
-                    if let Ok((after_expr, expr)) = expression(inner_start) {
-                        let after_expr = after_expr.trim_start();
-                        if let Some(after_paren) = after_expr.strip_prefix(')') {
-                            custom_traits.push((trait_name.clone(), Some(expr)));
-                            let (r2, _) = ws(after_paren)?;
-                            rest = r2;
-                            continue;
-                        }
-                    }
-                    let r2 = skip_balanced_parens(r);
-                    custom_traits.push((trait_name.clone(), None));
-                    let (r2, _) = ws(r2)?;
-                    rest = r2;
-                } else {
-                    // No parens — treat as parent role. If it's actually
-                    // a custom trait, the runtime will handle it via
-                    // trait_mod:<is> when the role name is not found.
-                    parent_roles.push(trait_name);
-                    let (r, _) = ws(r)?;
-                    rest = r;
-                }
-            }
-            continue;
-        }
-        if let Some(r) = keyword("hides", rest) {
-            let (r, _) = ws1(r)?;
-            let (r, hidden_name) = qualified_ident(r)?;
-            let (r, _) = ws(r)?;
-            // Track as a parent relationship
-            parent_roles.push(hidden_name.clone());
-            // Also mark the hidden relationship with a special marker
-            parent_roles.push(format!("__mutsu_role_hides__{}", hidden_name));
-            rest = r;
-            continue;
-        }
-        break;
-    }
-
-    let (rest, mut body) = match block(rest) {
-        Ok(ok) => ok,
-        Err(e) if e.is_fatal() => return Err(e),
-        Err(_) => consume_raw_braced_body(rest)?,
-    };
-    // Handle `also is rw;` in the role body
-    body.retain(|stmt| {
-        if stmt_is_also_is_rw(stmt) {
-            role_is_rw = true;
-            false
-        } else {
-            true
-        }
-    });
-    if is_hidden_role {
-        body.insert(
-            0,
-            Stmt::DoesDecl {
-                name: Symbol::intern("__mutsu_role_hidden__"),
-            },
-        );
-    }
-    for role_name in parent_roles.into_iter().rev() {
-        body.insert(
-            0,
-            Stmt::DoesDecl {
-                name: Symbol::intern(&role_name),
-            },
-        );
-    }
-    super::simple::register_user_type(&name);
-
-    let role_stmt = Stmt::RoleDecl {
-        name: Symbol::intern(&name),
-        type_params,
-        type_param_defs,
-        is_export,
-        export_tags,
-        body,
-        is_rw: role_is_rw,
-        language_version: super::simple::current_language_version(),
-        custom_traits,
-    };
-    // Emit __MUTSU_SET_META__ calls for ver/auth traits (like class declarations)
-    let mut meta_stmts = Vec::new();
-    for (trait_name, trait_value) in traits {
-        if trait_name == "ver" || trait_name == "auth" {
-            meta_stmts.push(meta_setter_stmt(&name, &trait_name, trait_value));
-        }
-    }
-    if meta_stmts.is_empty() {
-        return Ok((rest, role_stmt));
-    }
-    meta_stmts.push(role_stmt);
-    Ok((rest, Stmt::Block(meta_stmts)))
-}
-
-/// Parse `does` declaration.
-pub(super) fn does_decl(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("does", input).ok_or_else(|| PError::expected("does declaration"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, name) = parse_token_like_name(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (rest, _) = opt_char(rest, ';');
-    Ok((
-        rest,
-        Stmt::DoesDecl {
-            name: Symbol::intern(&name),
-        },
-    ))
-}
-
-/// Parse `trusts` declaration.
-pub(super) fn trusts_decl(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("trusts", input).ok_or_else(|| PError::expected("trusts declaration"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, name) = qualified_ident(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (rest, _) = opt_char(rest, ';');
-    Ok((
-        rest,
-        Stmt::TrustsDecl {
-            name: Symbol::intern(&name),
-        },
-    ))
-}
-
-/// Parse a `token`, `regex`, or `rule` declaration.
-pub(super) fn token_decl(input: &str) -> PResult<'_, Stmt> {
-    let is_rule = keyword("rule", input).is_some();
-    let is_regex = keyword("regex", input).is_some();
-    let is_ratchet = !is_regex; // token and rule are ratcheting
-    let rest = keyword("token", input)
-        .or_else(|| keyword("rule", input))
-        .or_else(|| keyword("regex", input))
-        .ok_or_else(|| PError::expected("token/regex/rule declaration"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, name) = parse_token_like_name(rest)?;
-    let (rest, _) = ws(rest)?;
-
-    // Optional params
-    let (rest, (params, param_defs)) = if rest.starts_with('(') {
-        let (r, _) = parse_char(rest, '(')?;
-        let (r, _) = ws(r)?;
-        let (r, pd) = parse_param_list(r)?;
-        let (r, _) = ws(r)?;
-        let (r, _) = parse_char(r, ')')?;
-        let names: Vec<String> = pd.iter().map(|p| p.name.clone()).collect();
-        (r, (names, pd))
-    } else {
-        (rest, (Vec::new(), Vec::new()))
-    };
-
-    let (rest, _) = ws(rest)?;
-    let (rest, mut pattern) = parse_raw_braced_regex_body(rest)?;
-    // An empty `token`/`regex`/`rule` body is a null regex.
-    if pattern.trim().is_empty() {
-        return Err(null_regex_error());
-    }
-    pattern = normalize_token_pattern(&pattern);
-    if is_rule {
-        pattern = inject_implicit_rule_ws(&pattern);
-        pattern = inject_separator_ws(&pattern);
-        if name.contains(":sym<") || name.contains(":sym\u{ab}") {
-            if !pattern.ends_with(' ') {
-                pattern.push(' ');
-            }
-            pattern.push_str("<.ws>?");
-        }
-    }
-    if is_ratchet {
-        pattern = format!(":ratchet {pattern}");
-    }
-    let body = vec![Stmt::Expr(Expr::Literal(Value::regex(pattern)))];
-
-    if is_rule {
-        Ok((
-            rest,
-            Stmt::RuleDecl {
-                name: Symbol::intern(&name),
-                params,
-                param_defs,
-                body,
-                multi: false,
-            },
-        ))
-    } else {
-        Ok((
-            rest,
-            Stmt::TokenDecl {
-                name: Symbol::intern(&name),
-                params,
-                param_defs,
-                body,
-                multi: false,
-                is_my: false,
-                is_our: false,
-            },
-        ))
-    }
-}
-
-/// Parse `grammar` declaration.
-pub(super) fn grammar_decl(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("grammar", input).ok_or_else(|| PError::expected("grammar declaration"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, name) = qualified_ident(rest)?;
-    check_pseudo_package_in_decl(&name)?;
-    // Consume optional type adverbs (`:ver<...>`, etc.) on the grammar name
-    // before the `is`/`does` parent clauses (e.g. `grammar Foo:ver<1> is Bar`).
-    let (rest, _traits) = parse_declarator_traits(rest)?;
-    let (rest, _) = ws(rest)?;
-    let mut r = rest;
-    let mut parents = Vec::new();
-    while let Some(r2) = keyword("is", r) {
-        let (r2, _) = ws1(r2)?;
-        let (r2, parent_name) = qualified_ident(r2)?;
-        parents.push(parent_name);
-        let (r2, _) = ws(r2)?;
-        r = r2;
-    }
-    let mut does_parents = Vec::new();
-    while let Some(r2) = keyword("does", r) {
-        let (r2, _) = ws1(r2)?;
-        let (r2, role_name) = qualified_ident(r2)?;
-        does_parents.push(role_name);
-        let (r2, _) = ws(r2)?;
-        let (r2, _) = skip_optional_role_args(r2)?;
-        r = r2;
-    }
-    // Default parent is Grammar if no `is` clause (unless the grammar itself is named Grammar)
-    if parents.is_empty() && name != "Grammar" {
-        parents.push("Grammar".to_string());
-    }
-    let (rest, body) = block(r)?;
-    super::simple::register_user_type(&name);
-    Ok((
-        rest,
-        Stmt::ClassDecl {
-            name: Symbol::intern(&name),
-            name_expr: None,
-            parents,
-            class_is_rw: false,
-            is_hidden: false,
-            is_lexical: false,
-            hidden_parents: vec![],
-            does_parents,
-            repr: None,
-            body,
-            language_version: super::simple::current_language_version(),
-            custom_traits: Vec::new(),
-            is_unit: false,
-        },
-    ))
-}
-
-/// Parse `module Name { ... }` declaration (non-unit form).
-pub(super) fn module_decl(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("module", input).ok_or_else(|| PError::expected("module declaration"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, name) = qualified_ident(rest)?;
-    check_pseudo_package_in_decl(&name)?;
-    let (rest, traits) = parse_declarator_traits(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (rest, body) = block(rest)?;
-    // Two `is export` declarations of the same symbol in one module clash.
-    if let Some(clash) = find_export_name_clash(&body) {
-        return Err(export_name_clash_error(&clash));
-    }
-    // Record exported subs from inline module so `import` can register them at parse time.
-    let exported = extract_exported_subs(&body);
-    if !exported.is_empty() {
-        super::simple::register_inline_module_exports(&name, exported);
-    }
-    let mut stmts = Vec::new();
-    for (trait_name, trait_value) in traits {
-        if trait_name == "ver" || trait_name == "auth" {
-            stmts.push(meta_setter_stmt(&name, &trait_name, trait_value));
-        }
-    }
-    let package_stmt = Stmt::Package {
-        name: Symbol::intern(&name),
-        body,
-        kind: crate::ast::PackageKind::Module,
-        is_unit: false,
-        is_my: false,
-    };
-    if stmts.is_empty() {
-        return Ok((rest, package_stmt));
-    }
-    stmts.push(package_stmt);
-    Ok((rest, Stmt::SyntheticBlock(stmts)))
-}
-
-/// Extract names of exported sub declarations from a statement list.
-fn extract_exported_subs(stmts: &[Stmt]) -> Vec<super::simple::InlineModuleExportSpec> {
-    let mut names = Vec::new();
-    for stmt in stmts {
-        match stmt {
-            Stmt::SubDecl {
-                name,
-                is_export,
-                precedence_trait,
-                associativity,
-                ..
-            } if *is_export => {
-                names.push((
-                    name.to_string(),
-                    precedence_trait.clone(),
-                    associativity.clone(),
-                ));
-            }
-            Stmt::SyntheticBlock(inner) => {
-                names.extend(extract_exported_subs(inner));
-            }
-            _ => {}
-        }
-    }
-    names
-}
-
-/// Recursively collect exported sub names (descending into nested blocks),
-/// returning the first symbol that is exported more than once. In Raku two
-/// `is export` declarations of the same symbol within one package raise
-/// X::Export::NameClash at compile time.
-fn find_export_name_clash(stmts: &[Stmt]) -> Option<String> {
-    let mut seen = std::collections::HashSet::new();
-    fn walk(stmts: &[Stmt], seen: &mut std::collections::HashSet<String>) -> Option<String> {
-        for stmt in stmts {
-            match stmt {
-                // `multi` candidates legitimately share a name, so only a
-                // non-multi (`only`) exported sub can clash.
-                Stmt::SubDecl {
-                    name,
-                    is_export,
-                    multi,
-                    ..
-                } if *is_export && !*multi => {
-                    if !seen.insert(name.to_string()) {
-                        return Some(name.to_string());
-                    }
-                }
-                Stmt::Block(inner) | Stmt::SyntheticBlock(inner) => {
-                    if let Some(clash) = walk(inner, seen) {
-                        return Some(clash);
-                    }
-                }
-                _ => {}
-            }
-        }
-        None
-    }
-    walk(stmts, &mut seen)
-}
-
-/// Build an X::Export::NameClash parse error for a symbol exported twice.
-fn export_name_clash_error(name: &str) -> PError {
-    let symbol = format!("&{}", name);
-    let msg = format!("A symbol '{}' has already been exported", symbol);
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("symbol".to_string(), Value::str(symbol));
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    let ex = Value::make_instance(Symbol::intern("X::Export::NameClash"), attrs);
-    PError::fatal_with_exception(msg, Box::new(ex))
-}
-
-/// Check if a declaration name is a pseudo-package and return an error if so.
-/// In Raku, the action is always "package name" regardless of the specific declarator.
-fn check_pseudo_package_in_decl(name: &str) -> Result<(), PError> {
-    if is_pseudo_package(name) {
-        let action = "package name";
-        let msg = format!("Cannot use pseudo package {} in {}", name, action);
-        let mut attrs = std::collections::HashMap::new();
-        attrs.insert("pseudo-package".to_string(), Value::str(name.to_string()));
-        attrs.insert("action".to_string(), Value::str(action.to_string()));
-        attrs.insert("message".to_string(), Value::str(msg.clone()));
-        let ex = Value::make_instance(Symbol::intern("X::PseudoPackage::InDeclaration"), attrs);
-        return Err(PError::fatal_with_exception(msg, Box::new(ex)));
-    }
-    Ok(())
-}
-
-/// Parse `unit module` or `unit class` statement.
-pub(super) fn unit_module_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("unit", input).ok_or_else(|| PError::expected("unit statement"))?;
-    let (rest, _) = ws1(rest)?;
-    // unit class Name;
-    // unit class Name is Parent does Role;
-    if let Some(r) = keyword("class", rest) {
-        let (r, _) = ws1(r)?;
-        let (r, name) = qualified_ident(r)?;
-        check_pseudo_package_in_decl(&name)?;
-        // Skip optional type adverbs (:ver<...>, :auth<...>, :api<...>)
-        let (r, _traits) = parse_declarator_traits(r)?;
-        let (r, _) = ws(r)?;
-        // Parse `is Parent` and `does Role` clauses before the semicolon
-        let mut parents = Vec::new();
-        let mut does_parents = Vec::new();
-        let mut class_is_rw = false;
-        let mut is_hidden = false;
-        let mut hidden_parents = Vec::new();
-        let mut r = r;
-        loop {
-            if let Some(r2) = keyword("is", r) {
-                let (r2, _) = ws1(r2)?;
-                let (r2, parent) = if let Some(stripped) = r2.strip_prefix("::") {
-                    let (r3, ident_part) = qualified_ident(stripped)?;
-                    (r3, format!("::{}", ident_part))
-                } else {
-                    qualified_ident(r2)?
-                };
-                if parent == "rw" {
-                    class_is_rw = true;
-                } else if parent == "hidden" {
-                    is_hidden = true;
-                } else {
-                    parents.push(parent);
-                }
-                let (r2, _) = ws(r2)?;
-                r = r2;
-                continue;
-            }
-            if let Some(r2) = keyword("does", r) {
-                let (r2, _) = ws1(r2)?;
-                let (r2, role_name) = qualified_ident(r2)?;
-                parents.push(role_name.clone());
-                does_parents.push(role_name);
-                let (r2, _) = ws(r2)?;
-                r = r2;
-                continue;
-            }
-            if let Some(r2) = keyword("hides", r) {
-                let (r2, _) = ws1(r2)?;
-                let (r2, parent) = qualified_ident(r2)?;
-                parents.push(parent.clone());
-                hidden_parents.push(parent);
-                let (r2, _) = ws(r2)?;
-                r = r2;
-                continue;
-            }
-            break;
-        }
-        let (r, _) = opt_char(r, ';');
-        return Ok((
-            r,
-            Stmt::ClassDecl {
-                name: Symbol::intern(&name),
-                name_expr: None,
-                parents,
-                class_is_rw,
-                is_hidden,
-                is_lexical: false,
-                hidden_parents,
-                does_parents,
-                repr: None,
-                body: Vec::new(),
-                language_version: super::simple::current_language_version(),
-                custom_traits: Vec::new(),
-                is_unit: true,
-            },
-        ));
-    }
-    // unit role Name;  — declare a role at the file scope.
-    if let Some(r) = keyword("role", rest) {
-        let (r, _) = ws1(r)?;
-        let (r, name) = qualified_ident(r)?;
-        check_pseudo_package_in_decl(&name)?;
-        let (r, _) = ws(r)?;
-        let (r, _) = opt_char(r, ';');
-        return Ok((
-            r,
-            Stmt::RoleDecl {
-                name: Symbol::intern(&name),
-                type_params: Vec::new(),
-                type_param_defs: Vec::new(),
-                is_export: false,
-                export_tags: Vec::new(),
-                body: Vec::new(),
-                is_rw: false,
-                language_version: super::simple::current_language_version(),
-                custom_traits: Vec::new(),
-            },
-        ));
-    }
-    // unit grammar Name;  — declare a grammar at the file scope.
-    // unit grammar Name is Parent;
-    if let Some(r) = keyword("grammar", rest) {
-        let (r, _) = ws1(r)?;
-        let (r, name) = qualified_ident(r)?;
-        check_pseudo_package_in_decl(&name)?;
-        // Consume optional type adverbs (`:ver<...>`, `:auth<...>`, `:api<...>`)
-        // on the grammar name before the `is`/`does` parent clauses, e.g.
-        // `unit grammar Foo:ver<0.3.8> is Bar;`. Without this, the adverb blocks
-        // the `is Bar` parent from being parsed (parent silently dropped).
-        let (r, _traits) = parse_declarator_traits(r)?;
-        let (r, _) = ws(r)?;
-        let mut r = r;
-        let mut parents = Vec::new();
-        let mut does_parents = Vec::new();
-        loop {
-            if let Some(r2) = keyword("is", r) {
-                let (r2, _) = ws1(r2)?;
-                let (r2, parent) = if let Some(stripped) = r2.strip_prefix("::") {
-                    let (r3, ident_part) = qualified_ident(stripped)?;
-                    (r3, format!("::{}", ident_part))
-                } else {
-                    qualified_ident(r2)?
-                };
-                parents.push(parent);
-                let (r2, _) = ws(r2)?;
-                r = r2;
-                continue;
-            }
-            if let Some(r2) = keyword("does", r) {
-                let (r2, _) = ws1(r2)?;
-                let (r2, role_name) = qualified_ident(r2)?;
-                parents.push(role_name.clone());
-                does_parents.push(role_name);
-                let (r2, _) = ws(r2)?;
-                r = r2;
-                continue;
-            }
-            break;
-        }
-        // Default parent is Grammar if no `is` clause
-        if parents.is_empty() && name != "Grammar" {
-            parents.push("Grammar".to_string());
-        }
-        let (r, _) = opt_char(r, ';');
-        super::simple::register_user_type(&name);
-        return Ok((
-            r,
-            Stmt::ClassDecl {
-                name: Symbol::intern(&name),
-                name_expr: None,
-                parents,
-                class_is_rw: false,
-                is_hidden: false,
-                is_lexical: false,
-                hidden_parents: vec![],
-                does_parents,
-                repr: None,
-                body: Vec::new(),
-                language_version: super::simple::current_language_version(),
-                custom_traits: Vec::new(),
-                is_unit: true,
-            },
-        ));
-    }
-    // Accept both `unit module Foo;` and `unit package Foo;`
-    let (rest, kind) = if let Some(r) = keyword("module", rest) {
-        (r, crate::ast::PackageKind::Module)
-    } else if let Some(r) = keyword("package", rest) {
-        (r, crate::ast::PackageKind::Package)
-    } else {
-        return Err(PError::expected("'module' or 'package' after 'unit'"));
-    };
-    let (rest, _) = ws1(rest)?;
-    let (rest, name) = qualified_ident(rest)?;
-    check_pseudo_package_in_decl(&name)?;
-    // Consume optional type adverbs (:ver<...>, :auth<...>, :api<...>) on the
-    // unit package name, e.g. `unit module Foo:ver<0.0.12>:auth<zef:bar>;`.
-    let (rest, _traits) = parse_declarator_traits(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (rest, _) = opt_char(rest, ';');
-    Ok((
-        rest,
-        Stmt::Package {
-            name: Symbol::intern(&name),
-            body: Vec::new(),
-            kind,
-            is_unit: true,
-            is_my: false,
-        },
-    ))
-}
-
-/// Parse `package` declaration.
-pub(super) fn package_decl(input: &str) -> PResult<'_, Stmt> {
-    package_decl_with_scope(input, false)
-}
-
-/// Parse `my package` declaration (lexically scoped).
-pub(super) fn package_decl_my(input: &str) -> PResult<'_, Stmt> {
-    package_decl_with_scope(input, true)
-}
-
-fn package_decl_with_scope(input: &str, is_my: bool) -> PResult<'_, Stmt> {
-    let rest = keyword("package", input).ok_or_else(|| PError::expected("package declaration"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, name) = qualified_ident(rest)?;
-    check_pseudo_package_in_decl(&name)?;
-    let (rest, _traits) = parse_declarator_traits(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (rest, body) = block(rest)?;
-    Ok((
-        rest,
-        Stmt::Package {
-            name: Symbol::intern(&name),
-            body,
-            kind: crate::ast::PackageKind::Package,
-            is_unit: false,
-            is_my,
-        },
-    ))
-}
-
-/// Parse `proto` declaration.
-pub(super) fn proto_decl(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("proto", input).ok_or_else(|| PError::expected("proto declaration"))?;
-    let (rest, _) = ws1(rest)?;
-    // proto token | proto rule | proto sub | proto method
-    let _is_token = keyword("token", rest).is_some() || keyword("rule", rest).is_some();
-    let is_method = keyword("method", rest).is_some() || keyword("submethod", rest).is_some();
-    let rest = if let Some(r) = keyword("token", rest)
-        .or_else(|| keyword("rule", rest))
-        .or_else(|| keyword("sub", rest))
-        .or_else(|| keyword("submethod", rest))
-        .or_else(|| keyword("method", rest))
-    {
-        let (r, _) = ws1(r)?;
-        r
-    } else {
-        rest
-    };
-    let (rest, name) = parse_sub_name(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (rest, param_defs) = if rest.starts_with('(') {
-        let (r, _) = parse_char(rest, '(')?;
-        let (r, _) = ws(r)?;
-        let (r, pd) = parse_param_list(r)?;
-        let (r, _) = ws(r)?;
-        let (r, _) = parse_char(r, ')')?;
-        (r, pd)
-    } else {
-        (rest, Vec::new())
-    };
-    let params: Vec<String> = param_defs.iter().map(|p| p.name.clone()).collect();
-    let (rest, _) = ws(rest)?;
-    // Parse traits (is export, etc.)
-    let (rest, traits) = parse_sub_traits(rest)?;
-    let (rest, _) = ws(rest)?;
-    // May have body or just semicolon
-    let mut body = Vec::new();
-    if rest.starts_with('{') {
-        let (rest, parsed_body) = match block(rest) {
-            Ok(ok) => ok,
-            Err(_) => consume_raw_braced_body(rest)?,
-        };
-        body = parsed_body;
-        return Ok((
-            rest,
-            Stmt::ProtoDecl {
-                name: Symbol::intern(&name),
-                params,
-                param_defs,
-                body,
-                is_export: traits.is_export,
-                custom_traits: traits
-                    .custom_traits
-                    .iter()
-                    .map(|(n, _)| n.clone())
-                    .collect(),
-                is_method,
-            },
-        ));
-    }
-    let (rest, _) = opt_char(rest, ';');
-    Ok((
-        rest,
-        Stmt::ProtoDecl {
-            name: Symbol::intern(&name),
-            params,
-            param_defs,
-            body,
-            is_export: traits.is_export,
-            custom_traits: traits
-                .custom_traits
-                .iter()
-                .map(|(n, _)| n.clone())
-                .collect(),
-            is_method,
-        },
-    ))
-}
+// Public entry points — preserve each function's original visibility.
+pub(super) use class_decl::{also_trait_stmt, class_decl_body};
+pub(crate) use class_decl::{anon_class_decl, augment_class_decl, class_decl};
+pub(super) use grammar_module::{does_decl, grammar_decl, module_decl, token_decl, trusts_decl};
+pub(super) use package_decl::{package_decl, package_decl_my, proto_decl, unit_module_stmt};
+pub(super) use role_decl::role_decl;
